Move middleware to its own module
